Extract deployment runtime into tau-deployment (#932)
