+++ crates/tau-coding-agent/Cargo.toml
@@ -27,6 +27,7 @@ tau-access = { path = "../tau-access" }
 tau-orchestrator = { path = "../tau-orchestrator" }
 tau-runtime = { path = "../tau-runtime" }
 tau-events = { path = "../tau-events" }
+tau-deployment = { path = "../tau-deployment" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
