+++ crates/tau-deployment/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "tau-deployment"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
+tau-core = { path = "../tau-core" }
+tau-runtime = { path = "../tau-runtime" }
+wasmparser.workspace = true
+tokio.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
