+++ crates/tau-deployment/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod deployment_contract;
+pub mod deployment_runtime;
+pub mod deployment_wasm;
